+++ packages/rallie/src/connector.ts
@@ -28,15 +28,15 @@ export class Connector<
   public methods: Methods
   public isRallieApp: boolean
 
-  public setState (setter: (state: State) => void | Promise<void>) {
+  public setState (action: string, setter: (state: State) => void | Promise<void>) {
     if (this.socket.existState(constant.stateNamespace(this.name))) {
-      return this.socket.setState(constant.stateNamespace(this.name), setter)
+      return this.socket.setState(constant.stateNamespace(this.name), action, setter)
     } else {
       throw new Error(errors.stateNotInitialized(this.name))
     }
   }
 
-  public watchState<P = any> (getter: (state: State, isWatchingEffect?: boolean) => undefined | P) {
+  public watchState<P = any> (getter: (state: State) => undefined | P) {
     if (this.socket.existState(constant.stateNamespace(this.name))) {
       return this.socket.watchState<State, P>(constant.stateNamespace(this.name), getter)
     } else {