+++ packages/react/test/apps/consumer.tsx
@@ -22,7 +22,7 @@ export const Consumer = () => {
     producer.methods.toggleTheme()
   }
   const addCount = () => {
-    producer.setState(state => {
+    producer.setState('add the count', state => {
       state.count++
     })
   }