+++ packages/react/test/apps/producer.tsx
@@ -28,7 +28,7 @@ export const Producer = () => {
   })
   useProducerMethods({
     toggleTheme () {
-      producer.setState(state => {
+      producer.setState('toggle theme', state => {
         state.isDarkTheme = !state.isDarkTheme
       })
     }